requirements.txt
fastapi==0.110.3
uvicorn==0.29.0
crewai==0.130.0
crewai-tools==0.47.1
langchain==0.1.52
langchain-community
langchain-openai
python-dotenv
requests
